pyproject.toml
[project]
name = "awww"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.3",
    "django-allauth>=65.16.1",
    "django-debug-toolbar>=6.3.0",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
]
